Persistence stage 2.
IMPORTANT NOTE: Raft is still semi-broken right now.
You'll be fine if you "rm ?/log*" before running your servers.

The followers now persist log entries.  It wasn't clear to me *when*
in the process that should happen, but given "Figure 8", it seems like
followers should only persist when they know an entry is committed.

The leader (and followers) now write multiple numbered log files, and
there's now a max file size for the log.  If the log wants to
persist an entry, and that would take it over the max log file size,
it increments to the next log file to write the entry.  (There's a
little needlessly-complicated logic that says, if we just opened this
log file, always write at least one entry.)

I think I only have the easy stuff left--persisting the two
integers as called for by raft.pdf figure 2.