Re-abstracted server, after persistence undid it.
The persistence changes had the Server and Log reading
and writing data to disk themselves!  With their *bare hands*.

This was acceptable only temporarily, as it meant that the
Server and Log now had their own independent state beyond
the control of the Driver.  In particular, this meant the
test driver couldn't control all the initial conditions,
nor test all the outputs.

The Driver now has four new pure virtual APIs:
    load_log
    load_state
    save_log
    save_state

The real implementation of these now live in the asyncio
driver.  The test driver now has toy versions of these.

Unfortunately, I broke the test driver, and it's late
and I'm not fixing it now.  That'll wait for another day.